Restructure webhook tests
