+++ conftest.py
@@ -34,8 +34,8 @@ pytest_plugins = [
     "saleor.graphql.order.tests.benchmark.fixtures",
     "saleor.graphql.giftcard.tests.benchmark.fixtures",
     "saleor.graphql.webhook.tests.benchmark.fixtures",
-    "saleor.plugins.webhook.tests.subscription_webhooks.fixtures",
     "saleor.tax.tests.fixtures",
+    "saleor.webhook.tests.subscription_webhooks.fixtures",
 ]
