--- saleor/plugins/webhook/tests/cassettes/test_trigger_webhooks_with_http.yaml
@@ -1,71 +1,0 @@
-interactions:
-- request:
-    body: '[{"model": "order.order", "pk": 187, "fields": {"private_metadata": {},
-      "metadata": {}, "created": "2020-08-07T09:05:59.192Z", "status": "unfulfilled",
-      "user": 202, "language_code": "en", "tracking_client_id": "", "billing_address":
-      825, "shipping_address": 826, "user_email": "test@example.com", "currency":
-      "USD", "shipping_method": 235, "shipping_method_name": "DHL", "shipping_price_net_amount":
-      "10.00", "shipping_price_gross_amount": "12.30", "token": "aad0c768-c6d8-41ee-a25f-73b7538941a8",
-      "checkout_token": "", "total_net_amount": "80.00", "total_gross_amount": "98.40",
-      "voucher": null, "discount_amount": "0.00", "discount_name": null, "translated_discount_name":
-      null, "display_gross_prices": true, "customer_note": "", "weight": "0.0:kg",
-      "gift_cards": []}}]'
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '771'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - python-requests/2.24.0
-
-      # X- headers will be deprecated in Saleor 4.0, proper headers are without X-
-      X-Saleor-Domain:
-      - mirumee.com
-      X-Saleor-Event:
-      - order_created
-      X-Saleor-Signature:
-      - ''
-      Saleor-Domain:
-      - mirumee.com
-      Saleor-Event:
-      - order_created
-      Saleor-Signature:
-      - ''
-    method: POST
-    uri: https://webhook.site/48978b64-4efb-43d5-a334-451a1d164009
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAAwMAAAAAAAAAAAA=
-    headers:
-      Cache-Control:
-      - no-cache, private
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - text/plain; charset=UTF-8
-      Date:
-      - Fri, 07 Aug 2020 09:05:59 GMT
-      Server:
-      - nginx/1.14.2
-      Set-Cookie:
-      - laravel_session=ylfogNz3uixEXKaOfz1tmSob8nISj4pH8vFG5XRc; expires=Fri, 07-Aug-2020
-        11:05:59 GMT; Max-Age=7200; path=/; httponly
-      Transfer-Encoding:
-      - chunked
-      Vary:
-      - Accept-Encoding
-      X-Request-Id:
-      - 0e944b79-e2e2-4f52-9879-1a9635416fb9
-      X-Token-Id:
-      - 48978b64-4efb-43d5-a334-451a1d164009
-    status:
-      code: 200
-      message: OK
-version: 1
