--- saleor/plugins/webhook/tests/cassettes/test_trigger_webhooks_with_http_and_secret_key.yaml
@@ -1,75 +1,0 @@
-interactions:
-- request:
-    body: '[{"model": "order.order", "pk": 195, "fields": {"private_metadata": {},
-      "metadata": {}, "created": "2020-08-07T09:06:01.255Z", "status": "unfulfilled",
-      "user": 210, "language_code": "en", "tracking_client_id": "", "billing_address":
-      857, "shipping_address": 858, "user_email": "test@example.com", "currency":
-      "USD", "shipping_method": 243, "shipping_method_name": "DHL", "shipping_price_net_amount":
-      "10.00", "shipping_price_gross_amount": "12.30", "token": "ee6320cf-3080-4578-bab2-520e93b4b065",
-      "checkout_token": "", "total_net_amount": "80.00", "total_gross_amount": "98.40",
-      "voucher": null, "discount_amount": "0.00", "discount_name": null, "translated_discount_name":
-      null, "display_gross_prices": true, "customer_note": "", "weight": "0.0:kg",
-      "gift_cards": []}}]'
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '771'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - python-requests/2.24.0
-
-      # X- headers will be deprecated in Saleor 4.0, proper headers are without X-
-      X-Saleor-Domain:
-      - mirumee.com
-      X-Saleor-Event:
-      - order_created
-      X-Saleor-HMAC-SHA256:
-      - sha1=cce609bb36a9aa6b54af010a2aef40979ddd56a309590a470a9c4602226324b4
-      X-Saleor-Signature:
-      - cce609bb36a9aa6b54af010a2aef40979ddd56a309590a470a9c4602226324b4
-      Saleor-Domain:
-      - mirumee.com
-      Saleor-Event:
-      - order_created
-      Saleor-HMAC-SHA256:
-      - sha1=cce609bb36a9aa6b54af010a2aef40979ddd56a309590a470a9c4602226324b4
-      Saleor-Signature:
-      - cce609bb36a9aa6b54af010a2aef40979ddd56a309590a470a9c4602226324b4
-    method: POST
-    uri: https://webhook.site/48978b64-4efb-43d5-a334-451a1d164009
-  response:
-    body:
-      string: !!binary |
-        H4sIAAAAAAAAAwMAAAAAAAAAAAA=
-    headers:
-      Cache-Control:
-      - no-cache, private
-      Content-Encoding:
-      - gzip
-      Content-Type:
-      - text/plain; charset=UTF-8
-      Date:
-      - Fri, 07 Aug 2020 09:06:01 GMT
-      Server:
-      - nginx/1.14.2
-      Set-Cookie:
-      - laravel_session=RJ3x6kNdLBbDUIJXj2TAYiiIlfP2pV0jBXWtUGjM; expires=Fri, 07-Aug-2020
-        11:06:01 GMT; Max-Age=7200; path=/; httponly
-      Transfer-Encoding:
-      - chunked
-      Vary:
-      - Accept-Encoding
-      X-Request-Id:
-      - 1d00628d-31d9-4636-b6b8-af7e35e3dde9
-      X-Token-Id:
-      - 48978b64-4efb-43d5-a334-451a1d164009
-    status:
-      code: 200
-      message: OK
-version: 1
