+++ saleor/webhook/tests/subscription_webhooks/test_create_deliveries_for_list_payment_methods.py
@@ -2,9 +2,9 @@ import json
 
 import graphene
 
-from .....payment.interface import ListStoredPaymentMethodsRequestData
-from .....webhook.event_types import WebhookEventSyncType
-from .....webhook.transport.asynchronous.transport import (
+from ....payment.interface import ListStoredPaymentMethodsRequestData
+from ...event_types import WebhookEventSyncType
+from ...transport.asynchronous.transport import (
     create_deliveries_for_subscriptions,
 )
