+++ saleor/webhook/tests/subscription_webhooks/test_create_deliveries_for_payment_gateway_initialize_session.py
@@ -3,11 +3,9 @@ from decimal import Decimal
 
 import graphene
 
-from .....webhook.event_types import WebhookEventSyncType
-from .....webhook.models import Webhook
-from .....webhook.transport.asynchronous.transport import (
-    create_deliveries_for_subscriptions,
-)
+from ...event_types import WebhookEventSyncType
+from ...models import Webhook
+from ...transport.asynchronous import create_deliveries_for_subscriptions
 
 PAYMENT_GATEWAY_INITIALIZE_SESSION = """
 subscription {
