+++ saleor/webhook/tests/subscription_webhooks/test_create_deliveries_for_subscriptions_payments.py
@@ -2,10 +2,8 @@ import json
 
 import graphene
 
-from .....webhook.event_types import WebhookEventSyncType
-from .....webhook.transport.asynchronous.transport import (
-    create_deliveries_for_subscriptions,
-)
+from ...event_types import WebhookEventSyncType
+from ...transport.asynchronous import create_deliveries_for_subscriptions
 from .payloads import generate_payment_payload
