+++ saleor/webhook/tests/subscription_webhooks/test_create_deliveries_for_translation_subscription.py
@@ -2,10 +2,8 @@ import json
 
 import graphene
 
-from .....webhook.event_types import WebhookEventAsyncType
-from .....webhook.transport.asynchronous.transport import (
-    create_deliveries_for_subscriptions,
-)
+from ...event_types import WebhookEventAsyncType
+from ...transport.asynchronous.transport import create_deliveries_for_subscriptions
 
 
 def test_translation_created_product(
