+++ saleor/webhook/tests/test_webhook_ip_filtering.py
@@ -1,6 +1,6 @@
-from ....core.http_client import HTTPClient
-from ....core.models import EventDeliveryAttempt
-from ....webhook.transport.asynchronous.transport import trigger_webhooks_async
+from ...core.http_client import HTTPClient
+from ...core.models import EventDeliveryAttempt
+from ..transport.asynchronous import trigger_webhooks_async
 
 
 def test_rejects_private_ips(webhook, monkeypatch):
