+++ saleor/webhook/transport/synchronous/__init__.py
@@ -1,0 +1,11 @@
+from .transport import (
+    trigger_all_webhooks_sync,
+    trigger_webhook_sync,
+    trigger_webhook_sync_if_not_cached,
+)
+
+__all__ = [
+    "trigger_all_webhooks_sync",
+    "trigger_webhook_sync",
+    "trigger_webhook_sync_if_not_cached",
+]
